requirements.txt
amqp==5.0.6
asgiref==3.3.1
backoff==1.10.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
Django==3.1.3
django-analytical==3.0.0
django-bootstrap-pagination==1.7.1
django-countries==6.1.3
django-crispy-forms==1.9.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-resized==0.3.11
django-social-share==2.1.0
django-social-widgets==0.4.0
djangorestframework==3.12.2
geographiclib==1.50
geopy==2.0.0
haversine==2.3.0
idna==2.10
kombu==5.0.2
opencage==1.2.2
paystackapi==2.0.0
phonenumbers==8.12.12
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==19.1.0
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wincertstore==0.2
